Converted images to webp, beautified code, seperated SASS variables, modified folder striture and added comments.
