+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "library-app",
+  "name": "sr-library-app",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
@@ -7,10 +7,10 @@
     "@testing-library/react": "^11.2.6",
     "@testing-library/user-event": "^12.8.3",
     "@types/jest": "^26.0.23",
-    "@types/node": "^12.20.13",
+    "@types/node": "^12.20.12",
     "@types/react": "^17.0.5",
-    "@types/react-dom": "^17.0.5",
-    "bootstrap": "^5.0.1",
+    "@types/react-dom": "^17.0.4",
+    "bootstrap": "^4.6.0",
     "react": "^17.0.2",
     "react-bootstrap": "^1.6.0",
     "react-confirm-alert": "^2.7.0",
@@ -49,6 +49,7 @@
     ]
   },
   "devDependencies": {
+    "@types/node-sass": "^4.11.1",
     "@types/react-select": "^4.0.15"
   }
 }
