+++ src/App.tsx
@@ -1,7 +1,12 @@
 import React from 'react';
-import ClientApp from "./ClientApp";
+
+// Component imports
+import ClientApp from "./views/ClientApp";
+
+// Style imports
 import './app.scss';
 
+// Main component
 const App: React.FC = () => {
     return (
         <>
@@ -11,4 +16,3 @@ const App: React.FC = () => {
 };
 
 export default App;
-
