+++ src/app.scss
@@ -1,2 +1,4 @@
-@use "./main";
+@use "components/main";
+@import "~bootstrap/scss/bootstrap";
+
