+++ src/components/LibrarySection/BookSection/__bookSection.scss
@@ -1,12 +1,4 @@
-$header-color: #2e393d;
-$header-underline-color: #2b2f32;
-$list-color: #2e2e2e;
-$list-bg-hover-color: #f5f5f5;
-$form-button-color: #3f6cbb;
-$form-title-color: #074d43;
-$form-close-button-color: #323232;
-$form-input-color: #e0e0e0;
-$form-button-color: #01589d;
+@import "../../_variables";
 
 /* =============================
     Book Section CSS
