+++ src/components/__common.scss
@@ -1,0 +1,11 @@
+/* =============================
+    Common CSS
+================================ */
+
+body, html {
+  font-family: 'Raleway', sans-serif !important;
+}
+
+h1, h2, h3, h4, h5, h6 {
+  font-family: 'Raleway', sans-serif;
+}
