--- src/components/__layout.scss
@@ -1,3 +1,0 @@
-body, html {
-  font-family: 'Raleway', sans-serif !important;
-}
