+++ src/components/__variables.scss
@@ -1,0 +1,13 @@
+/* =============================
+    SCSS Variables
+================================ */
+
+$header-color: #2e393d;
+$header-underline-color: #2b2f32;
+$list-color: #2e2e2e;
+$list-bg-hover-color: #f5f5f5;
+$form-button-color: #3f6cbb;
+$form-title-color: #22584b;
+$form-close-button-color: #323232;
+$form-input-color: #2e2e2e;
+$form-button-color: #01589d;
