+++ src/components/main.scss
@@ -1,0 +1,6 @@
+@use "_common";
+@use "WelcomeSection/_welcomeSection";
+@use "FooterSection/_footerSection";
+@use "LibrarySection/AuthorSection/_authorSection";
+@use "LibrarySection/BookSection/_bookSection";
+
