+++ src/index.tsx
@@ -1,15 +1,19 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
+
+// Style imports
 import './index.css';
-import App from './App';
 import 'bootstrap/dist/css/bootstrap.min.css';
+
+// Component imports
+import App from './App';
 import reportWebVitals from './reportWebVitals';
 
 ReactDOM.render(
-  <React.StrictMode>
-    <App />
-  </React.StrictMode>,
-  document.getElementById('root')
+    <React.StrictMode>
+        <App/>
+    </React.StrictMode>,
+    document.getElementById('root')
 );
 
 // If you want to start measuring performance in your app, pass a function
