--- src/main.scss
@@ -1,4 +1,0 @@
-@use "./components/WelcomeSection/_welcomeSection";
-@use "./components/__layout";
-@use "components/LibrarySection/AuthorSection/_authorSection";
-
